requirements.txt
Django==6.0.3
reportlab==4.4.10
Pillow==12.1.1
polib==1.2.0